test: move nonce manager to starknet api test utils
