+++ crates/blockifier/src/concurrency/versioned_state_test.rs
@@ -11,6 +11,7 @@ use starknet_api::core::{
     Nonce,
     PatriciaKey,
 };
+use starknet_api::test_utils::NonceManager;
 use starknet_api::transaction::{Calldata, ContractAddressSalt, ValidResourceBounds};
 use starknet_api::{
     calldata,
@@ -46,7 +47,7 @@ use crate::test_utils::contracts::FeatureContract;
 use crate::test_utils::deploy_account::deploy_account_tx;
 use crate::test_utils::dict_state_reader::DictStateReader;
 use crate::test_utils::initial_test_state::test_state;
-use crate::test_utils::{CairoVersion, NonceManager, BALANCE, DEFAULT_STRK_L1_GAS_PRICE};
+use crate::test_utils::{CairoVersion, BALANCE, DEFAULT_STRK_L1_GAS_PRICE};
 use crate::transaction::account_transaction::AccountTransaction;
 use crate::transaction::objects::{HasRelatedFeeType, TransactionInfoCreator};
 use crate::transaction::test_utils::{l1_resource_bounds, max_resource_bounds};
