+++ crates/starknet_api/src/test_utils.rs
@@ -1,3 +1,30 @@
+use std::collections::HashMap;
+
+use starknet_types_core::felt::Felt;
+
+use crate::core::{ContractAddress, Nonce};
+
 pub mod declare;
 pub mod deploy_account;
 pub mod invoke;
+
+#[derive(Default)]
+pub struct NonceManager {
+    next_nonce: HashMap<ContractAddress, Felt>,
+}
+
+impl NonceManager {
+    pub fn next(&mut self, account_address: ContractAddress) -> Nonce {
+        let next = self.next_nonce.remove(&account_address).unwrap_or_default();
+        self.next_nonce.insert(account_address, next + 1);
+        Nonce(next)
+    }
+
+    /// Decrements the nonce of the account, unless it is zero.
+    pub fn rollback(&mut self, account_address: ContractAddress) {
+        let current = *self.next_nonce.get(&account_address).unwrap_or(&Felt::default());
+        if current != Felt::ZERO {
+            self.next_nonce.insert(account_address, current - 1);
+        }
+    }
+}
